#63 Refactor components
