+++ app/core/components/CardBox/CardBox.styles.tsx
@@ -1,0 +1,16 @@
+import styled from "@emotion/styled"
+
+export const CardBoxStyle = styled.div`
+  background-color: #fff;
+  padding: 30px;
+  border-radius: 7px;
+  height: 100%;
+  .CardBox--title {
+    color: #252a2f;
+    font-size: 22px;
+    font-weight: 700;
+  }
+  .CardBox--content {
+    margin-top: 20px;
+  }
+`
