+++ app/core/components/DropDownButton/DropDownButton.styles.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled"
+
+export const DropdownPlaceholderContainer = styled.div`
+  display: flex;
+  cursor: pointer;
+  color: ;
+`
