+++ app/core/components/LabeledTextAreaField/LabeledTextAreaField.styles.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled"
+
+export const LabelComponent = styled.label`
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  font-size: 1rem;
+`
