+++ app/pages/projects/[projectId]/[projectId].styles.tsx
@@ -64,3 +64,4 @@ export const DetailMoreHead = styled.div`
     display: inline-block;
   }
 `
+export default HeaderInfo
