+++ app/pages/projects/full/full.styles.tsx
@@ -14,3 +14,4 @@ export const WrapperContentForm = styled.div`
   width: 100%;
   max-width: 540px;
 `
+export default WrapperContent
